Move functions to manage values in nested dict to utils
